# swagger/appload.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Appload Trading Room Application
  # license:
  #   name: MIT
servers:
  - url: http://www.appload.com/trading
paths:
...
  /product/publish:
    post:
      summary: Publish product for approval
      operationId: publishProduct
      tags:
        - product
      requestBody:
        description: Product object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: A list of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
  /product/approve/{id}:
    post:
      summary: Approve a product publish request
      operationId: approveProduct
      tags:
        - product
      parameters:
        - name: id
          in: path
          required: true
          description: The ID for the product to be approved
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: The product object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: The ID provided does not exist
        default:
          description: unexpected error
  /product/purchase/{id}:
    post:
      summary: Request a product for purchase
      operationId: purchaseProduct
      tags:
        - product
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the requested product
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: The product object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: The ID provided does not exist
        default:
          description: unexpected error
  /product/edit/{id}:
    post:
      summary: Edit a product of the given ID
      operationId: editProduct
      tags:
        - product
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the requested product
          schema:
            $ref: "#/components/schemas/Id"
      requestBody:
        description: Product object with the new data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: The product object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: The ID provided does not exist
        default:
          description: unexpected error
...
components:
  schemas:
...
    Product:
      type: object
      #required
      properties:
        id:
          $ref: "#/components/schemas/Id"
        userId:
          $ref: "#/components/schemas/Id"
        name:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
          format: double
        approved:
          type: boolean
        sold:
          type: boolean
...
    ProductCategory:
      type: string
      enum:
        - electronics
        - furniture
        - cosmetics
        - art
        - clothes
...
    Id:
      type: string
      format: uuid